+++ tests/image_generation/helpers/image_generation_controlnet_test_helper.py
@@ -1,12 +1,11 @@
 import os
 
-import numpy as np
-from PIL import Image
-
 from mflux.config.config import Config
 from mflux.config.model_config import ModelConfig
 from mflux.controlnet.flux_controlnet import Flux1Controlnet
 from tests.image_generation.helpers.image_generation_test_helper import ImageGeneratorTestHelper
+
+from .image_compare import check_images_close_enough
 
 
 class ImageGeneratorControlnetTestHelper:
@@ -55,13 +54,13 @@ class ImageGeneratorControlnetTestHelper:
             image.save(path=output_image_path, overwrite=True)
 
             # then
-            np.testing.assert_array_equal(
-                np.array(Image.open(output_image_path)),
-                np.array(Image.open(reference_image_path)),
-                err_msg="Generated image doesn't match reference image",
+            check_images_close_enough(
+                output_image_path,
+                reference_image_path,
+                "Generated controlnet image doesn't match reference image",
             )
 
         finally:
             # cleanup
-            if os.path.exists(output_image_path):
+            if os.path.exists(output_image_path) and "MFLUX_PRESERVE_TEST_OUTPUT" not in os.environ:
                 os.remove(output_image_path)
